# perf: use `capacity_builder::FastDisplay`