--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,7 +16,13 @@ serde = { version = "1.0.130", features = ["derive", "rc"] }
 thiserror = "2"
 url = "2.5.3"
 deno_error = "0.5.1"
+capacity_builder = "0.2.0"
 
 [dev-dependencies]
+divan = "0.1.17"
 pretty_assertions = "1.0.0"
 serde_json = { version = "1.0.67", features = ["preserve_order"] }
+
+[[bench]]
+name = "bench"
+harness = false