--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,7 +16,7 @@ serde = { version = "1.0.130", features = ["derive", "rc"] }
 thiserror = "2"
 url = "2.5.3"
 deno_error = "0.5.1"
-capacity_builder = "0.3.0"
+capacity_builder = "0.4.0"
 
 [dev-dependencies]
 divan = "0.1.17"