--- a/src/package.rs
+++ b/src/package.rs
@@ -1,5 +1,8 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
+use capacity_builder::FastDisplay;
+use capacity_builder::StringBuildable;
+use capacity_builder::StringBuilder;
 use deno_error::JsError;
 use monch::ParseErrorFailure;
 use serde::Deserialize;
@@ -70,10 +73,22 @@ pub struct PackageReqReferenceInvalidWithVersionParseError {
 ///
 /// This contains all the information found in a package specifier other than
 /// what kind of package specifier it was.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, FastDisplay)]
 pub struct PackageReqReference {
   pub req: PackageReq,
   pub sub_path: Option<String>,
+}
+
+impl StringBuildable for PackageReqReference {
+  fn string_build_with<'a>(&'a self, builder: &mut StringBuilder<'a, '_, '_>) {
+    if let Some(sub_path) = &self.sub_path {
+      builder.append(&self.req);
+      builder.append('/');
+      builder.append(sub_path);
+    } else {
+      builder.append(&self.req);
+    }
+  }
 }
 
 impl PackageReqReference {
@@ -125,16 +140,6 @@ impl PackageReqReference {
   }
 }
 
-impl std::fmt::Display for PackageReqReference {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    if let Some(sub_path) = &self.sub_path {
-      write!(f, "{}/{}", self.req, sub_path)
-    } else {
-      write!(f, "{}", self.req)
-    }
-  }
-}
-
 #[derive(Error, Debug, Clone, JsError)]
 pub enum PackageReqPartsParseError {
   #[class(type)]
@@ -166,19 +171,21 @@ pub struct PackageReqParseError {
 }
 
 /// The name and version constraint component of an `PackageReqReference`.
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, FastDisplay)]
 pub struct PackageReq {
   pub name: String,
   pub version_req: VersionReq,
 }
 
-impl std::fmt::Display for PackageReq {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl StringBuildable for PackageReq {
+  fn string_build_with<'a>(&'a self, builder: &mut StringBuilder<'a, '_, '_>) {
     if self.version_req.version_text() == "*" {
       // do not write out the version requirement when it's the wildcard version
-      write!(f, "{}", self.name)
+      builder.append(&self.name);
     } else {
-      write!(f, "{}@{}", self.name, self.version_req)
+      builder.append(&self.name);
+      builder.append('@');
+      builder.append(&self.version_req.raw_text);
     }
   }
 }
@@ -303,7 +310,12 @@ impl PackageReq {
 
   /// Outputs a normalized string representation of the package requirement.
   pub fn to_string_normalized(&self) -> String {
-    format!("{}@{}", self.name, self.version_req.inner())
+    StringBuilder::build(|builder| {
+      builder.append(&self.name);
+      builder.append('@');
+      builder.append(self.version_req.inner());
+    })
+    .unwrap()
   }
 }
 
@@ -458,7 +470,7 @@ pub struct PackageNvReferenceParseError {
 }
 
 /// A package name and version with a potential subpath.
-#[derive(Debug, Clone, PartialOrd, Ord, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialOrd, Ord, PartialEq, Eq, Hash, FastDisplay)]
 pub struct PackageNvReference {
   pub nv: PackageNv,
   pub sub_path: Option<String>,
@@ -526,12 +538,14 @@ impl PackageNvReference {
   }
 }
 
-impl std::fmt::Display for PackageNvReference {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl StringBuildable for PackageNvReference {
+  fn string_build_with<'a>(&'a self, builder: &mut StringBuilder<'a, '_, '_>) {
     if let Some(sub_path) = &self.sub_path {
-      write!(f, "{}/{}", self.nv, sub_path)
+      builder.append(&self.nv);
+      builder.append('/');
+      builder.append(sub_path);
     } else {
-      write!(f, "{}", self.nv)
+      builder.append(&self.nv);
     }
   }
 }
@@ -544,7 +558,7 @@ pub struct PackageNvParseError {
   pub text: String,
 }
 
-#[derive(Clone, PartialOrd, Ord, PartialEq, Eq, Hash)]
+#[derive(Clone, PartialOrd, Ord, PartialEq, Eq, Hash, FastDisplay)]
 pub struct PackageNv {
   pub name: String,
   pub version: Version,
@@ -557,9 +571,11 @@ impl std::fmt::Debug for PackageNv {
   }
 }
 
-impl std::fmt::Display for PackageNv {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    write!(f, "{}@{}", self.name, self.version)
+impl StringBuildable for PackageNv {
+  fn string_build_with<'a>(&'a self, builder: &mut StringBuilder<'a, '_, '_>) {
+    builder.append(&self.name);
+    builder.append('@');
+    builder.append(&self.version);
   }
 }