Before:

```
Timer precision: 100 ns
bench                       fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ package_req                            │               │               │               │         │
│  ╰─ to_string_normalized  1.296 µs      │ 1.491 µs      │ 1.365 µs      │ 1.37 µs       │ 100     │ 100000
├─ version                                │               │               │               │         │
│  ├─ to_string             239.5 ns      │ 316.3 ns      │ 244.5 ns      │ 247.4 ns      │ 100     │ 100000
│  ╰─ to_string_display     244.1 ns      │ 381.8 ns      │ 250.4 ns      │ 256.8 ns      │ 100     │ 100000
╰─ version_req                            │               │               │               │         │
   ├─ to_string             1.101 µs      │ 1.464 µs      │ 1.133 µs      │ 1.141 µs      │ 100     │ 100000
   ├─ to_string_display     1.109 µs      │ 1.3 µs        │ 1.135 µs      │ 1.142 µs      │ 100     │ 100000
   ╰─ to_string_normalized  1.185 µs      │ 1.474 µs      │ 1.232 µs      │ 1.244 µs      │ 100     │ 100000
```

After:

```
Timer precision: 100 ns
bench                       fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ package_req                            │               │               │               │         │
│  ╰─ to_string_normalized  1.273 µs      │ 1.459 µs      │ 1.317 µs      │ 1.327 µs      │ 100     │ 100000
├─ version                                │               │               │               │         │
│  ├─ to_string             183.8 ns      │ 268.3 ns      │ 192.3 ns      │ 192.4 ns      │ 100     │ 100000
│  ╰─ to_string_display     234.3 ns      │ 384.9 ns      │ 243.1 ns      │ 246.9 ns      │ 100     │ 100000
╰─ version_req                            │               │               │               │         │
   ├─ to_string             1.086 µs      │ 1.352 µs      │ 1.123 µs      │ 1.133 µs      │ 100     │ 100000
   ├─ to_string_display     1.102 µs      │ 1.371 µs      │ 1.14 µs       │ 1.149 µs      │ 100     │ 100000
   ╰─ to_string_normalized  1.151 µs      │ 1.289 µs      │ 1.169 µs      │ 1.179 µs      │ 100     │ 100000
```